+++ internal/consumer/types_test.go
@@ -24,6 +24,7 @@ import (
 	"sync/atomic"
 	"testing"
 
+	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -61,4 +62,29 @@ func TestSaramaClient(t *testing.T) {
 	// Second close should return no error and not increment closed counter
 	assert.NoError(t, c.Close())
 	assert.EqualValues(t, 1, atomic.LoadInt32(&mock.closed))
+}
+
+func TestProtobufDLQMetadataDecoder(t *testing.T) {
+	dlqMetadata := newDLQMetadata()
+	b, err := proto.Marshal(dlqMetadata)
+	assert.NoError(t, err)
+	decodedDLQMetadata, err := ProtobufDLQMetadataDecoder(b)
+	assert.NoError(t, err)
+	assert.EqualValues(t, dlqMetadata, decodedDLQMetadata)
+
+	_, err = ProtobufDLQMetadataDecoder(nil)
+	assert.Error(t, err)
+
+	_, err = ProtobufDLQMetadataDecoder([]byte{1, 2, 3, 4})
+	assert.Error(t, err)
+}
+
+func TestNoopDLQMetadataDecoder(t *testing.T) {
+	dlqMetadata := newDLQMetadata()
+	dlqMetadata.Offset = 100
+	b, err := proto.Marshal(dlqMetadata)
+	assert.NoError(t, err)
+	decodedDLQMetadata, err := NoopDLQMetadataDecoder(b)
+	assert.NoError(t, err)
+	assert.EqualValues(t, newDLQMetadata(), decodedDLQMetadata)
 }